+++ tests/test_core_engine/test_backends/test_readers/test_csv.py
@@ -5,13 +5,20 @@
 import csv
 from pathlib import Path
 from typing import Dict, Iterator, Optional
+from uuid import uuid4
 
 import pandas as pd
 import pytest
 from pydantic import BaseModel
 
-from dve.core_engine.backends.exceptions import EmptyFileError, FieldCountMismatch
+from dve.core_engine.backends.exceptions import (
+    EmptyFileError,
+    FieldCountMismatch,
+    MessageBearingError,
+    UnableToParseCSVError,
+)
 from dve.core_engine.backends.readers import CSVFileReader
+from dve.core_engine.backends.readers.utilities import get_all_model_fields
 from dve.core_engine.constants import RECORD_INDEX_COLUMN_NAME
 
 from ....conftest import get_test_file_path
@@ -22,6 +29,12 @@ from ....fixtures import temp_dir
 def planet_location() -> Iterator[str]:
     """The URI of the planet data"""
     yield get_test_file_path("planets/planets.csv").as_uri()
+
+
+@pytest.fixture(scope="module")
+def planet_additional_field_location() -> Iterator[str]:
+    """The URI for planet data with additional columns"""
+    yield get_test_file_path("planets/planets_add_fields.csv").as_uri()
 
 
 @pytest.fixture(scope="function")
@@ -102,6 +115,12 @@ class Planets(PlanetsSubset):
     number_of_moons: str
     has_ring_system: str
     has_global_magnetic_field: str
+
+
+class PlanetsWithExtra(Planets):
+    """A subset of the planets data with an extra field."""
+
+    random_null: str
 
 
 class SingleColumnModel(BaseModel):
@@ -261,3 +280,40 @@ class TestParametrizedCSVParser:
             pd.read_parquet(target_location).to_dict(orient="records"),
             key=lambda x: x.get(RECORD_INDEX_COLUMN_NAME),
         ) == sorted([dict(val) for val in planet_data.values()], key=lambda x: x.get(RECORD_INDEX_COLUMN_NAME))
+
+    def test_base_csv_reader_with_additional_fields(
+        self,
+        planet_additional_field_location: str
+    ):
+        """Test that message bearing error raised when additional fields provided"""
+        reader = CSVFileReader(field_check=True)
+        with pytest.raises(MessageBearingError) as exc_info:
+            list(reader.read_to_py_iterator(
+                planet_additional_field_location,
+                "test",
+                Planets,
+                get_all_model_fields([Planets])
+            ))
+
+        error_msg = exc_info.value.messages[0]
+        assert error_msg.record["test"] == "additional fields: add_field1, add_field2;"
+        assert "missing_fields" not in error_msg.record["test"]
+
+    def test_base_csv_reader_with_missing_fields(
+        self,
+        planet_location: str
+    ):
+        """Test that message bearing error raised when fields are missing from the expected schema"""
+        reader = CSVFileReader(field_check=True)
+
+        with pytest.raises(MessageBearingError) as exc_info:
+            list(reader.read_to_py_iterator(
+                planet_location,
+                "test",
+                PlanetsWithExtra,
+                get_all_model_fields([PlanetsWithExtra])
+            ))
+
+        error_msg = exc_info.value.messages[0]
+        assert "additional_fields" not in error_msg.record["test"]
+        assert error_msg.record["test"] == "missing fields: random_null;"