+++ tests/test_core_engine/test_backends/test_readers/test_utilities.py
@@ -1,55 +1,24 @@
-import datetime as dt
-from pathlib import Path
-import tempfile
-from uuid import uuid4
+"""Test utility functions & objects for readers"""
 
-import pytest
-from pydantic import BaseModel, create_model
+from pydantic import BaseModel
 
-from dve.core_engine.backends.readers.utilities import check_csv_header_expected
+from dve.core_engine.backends.readers.utilities import get_all_model_fields
 
-@pytest.mark.parametrize(
-    ["header_row", "delim", "schema", "expected"],
-    [
-        (
-            "field1,field2,field3",
-            ",",
-            {"field1": (str, ...), "field2": (int, ...), "field3": (float, 1.2)},
-            set(),
-        ),
-        (
-            "field2,field3,field1",
-            ",",
-            {"field1": (str, ...), "field2": (int, ...), "field3": (float, 1.2)},
-            set(),
-        ),
-        (
-            "str_field|int_field|date_field|",
-            ",",
-            {"str_field": (str, ...), "int_field": (int, ...), "date_field": (dt.date, dt.date.today())},
-            {"str_field","int_field","date_field"},
-        ),
-        (
-            '"str_field"|"int_field"|"date_field"',
-            "|",
-            {"str_field": (str, ...), "int_field": (int, ...), "date_field": (dt.date, dt.date.today())},
-            set(),
-        ),
-        (
-            'str_field,int_field,date_field\n',
-            ",",
-            {"str_field": (str, ...), "int_field": (int, ...), "date_field": (dt.date, dt.date.today())},
-            set(),
-        ),
-        
-    ],
-)
-def test_check_csv_header_expected(
-    header_row: str, delim: str, schema: type[BaseModel], expected: set[str]
-):
-    mdl = create_model("TestModel", **schema)
-    with tempfile.TemporaryDirectory() as tmpdir:
-        fle = Path(tmpdir).joinpath(f"test_file_{uuid4().hex}.csv")
-        fle.open("w+").write(header_row)
-        res = check_csv_header_expected(fle.as_posix(), mdl, delim)
-    assert res == expected
+
+class Model1(BaseModel):  # pylint: disable=C0115
+    model1_field_1: str
+    model1_field_2: int
+
+
+class Model2(BaseModel):  # pylint: disable=C0115
+    model2_field_1: str
+
+
+def test_get_all_model_fields():
+    """Test get_all_model_fields returns a unique set of fields from multiple models"""
+    md1 = Model1(model1_field_1="hello", model1_field_2=123)
+    md2 = Model2(model2_field_1="world")
+
+    result = get_all_model_fields([md1, md2])
+
+    assert result == {"model1_field_1", "model1_field_2", "model2_field_1"}